Fix reviewer's comment
